+++ src/createRollupGetRetryablesFees.ts
@@ -13,6 +13,7 @@ import { getRollupCreatorAddress } from './utils/getRollupCreatorAddress';
 import { isCustomFeeTokenAddress } from './utils/isCustomFeeTokenAddress';
 import { defaults as createRollupDefaults } from './createRollupDefaults';
 import { applyPercentIncrease } from './utils/gasOverrides';
+import { createRollupDefaultRetryablesFees } from './constants';
 
 const deployHelperABI = [
   {
@@ -115,4 +116,21 @@ export async function createRollupGetRetryablesFees<TChain extends Chain | undef
     functionName: 'getDeploymentTotalCost',
     data: result!,
   });
+}
+
+export async function createRollupGetRetryablesFeesWithFallback<TChain extends Chain | undefined>(
+  publicClient: PublicClient<Transport, TChain>,
+  { nativeToken, maxFeePerGasForRetryables }: CreateRollupGetRetryablesFeesParams,
+): Promise<bigint> {
+  try {
+    return await createRollupGetRetryablesFees(publicClient, {
+      nativeToken,
+      maxFeePerGasForRetryables,
+    });
+  } catch (error) {
+    console.error(
+      '[createRollupGetRetryablesFeesWithFallback] failed to fetch retryables fees, falling back to defaults.',
+    );
+    return createRollupDefaultRetryablesFees;
+  }
 }
